Category restore, previews
Re-opening the interface now synchronizes the script with the UI in a
hacky way. The dimensions of an invisible image are manipulated to
represent the current category selection state and the selected category
index (see sip.init and sip.selectCategory)
This can be reworked once CF fixes the calback widget.getSelectedOption.

Big previews currently use the inventory icon. Should add a function to
fetch a big preview image (object orientation, item largeImage, etc.).